packages/web/src/member-eligibility/sign-up/screens/delivery-address-screen/amplitude-events/DeliveryAddress.ts
import { EligibilityDetails } from '@/root/src/member-eligibility/shared/hooks/use-eligibility-details/types/eligibliity-details/EligibilityDetails';
import { AmplitudeEvents } from '@/root/src/member-eligibility/shared/amplitude-events/AmplitudeEvents';

export const deliveryAddressEvents: AmplitudeEvents = {
  onBackClicked: (eligibilityDetails: EligibilityDetails) => ({
    event: eligibilityDetails.flow === 'Sign Up' ? 'signup_click' : 'renewal_click',
    params: {
      page_name: 'DeliveryAddress',
      CTA: 'back',
    },
  }),
  onFinishClicked: (eligibilityDetails: EligibilityDetails) => ({
    event: eligibilityDetails.flow === 'Sign Up' ? 'signup_click' : 'renewal_click',
    params: {
      page_name: 'DeliveryAddress',
      CTA: 'finish',
    },
  }),
};
